+++ NEWS.md
@@ -10,6 +10,7 @@
 ### Features
 * Create feature flag for enabling optimization to prevent redundant updates in Inventory (Instance,Holding,Item) ([MODINVSTOR-1577](https://folio-org.atlassian.net/browse/MODINVSTOR-1577))
 * Create instance custom link setting ([MODINVSTOR-1600](https://folio-org.atlassian.net/browse/MODINVSTOR-1600))
+* Add `includeShadowCopies` query parameter to `GET /instance-storage/instances` ([MODINVSTOR-1608](https://folio-org.atlassian.net/browse/MODINVSTOR-1608))
 
 ### Bug fixes
 * Populate item and holdings `hrId` in the items-and-holdings view ([MODINVSTOR-1587](https://folio-org.atlassian.net/browse/MODINVSTOR-1587))